Refactor database control
