+++ birdsi/__init__.py
@@ -1,0 +1,3 @@
+from flask import Flask, session, redirect, request, url_for, \
+	render_template, g
+from sqlite3 import dbapi2 as sqlite3
