--- birdsi/models.py
@@ -1,17 +1,0 @@
-import sqlite3 as sql
-
-def insertUser(username,password):
-    con = sql.connect("birdsi.db")
-    cur = con.cursor()
-    cur.execute("INSERT INTO users (username,password) VALUES (?,?)", (username,password))
-    con.commit()
-    con.close()
-
-def retrieveUsers():
-    con = sql.connect("birdsi.db")
-    cur = con.cursor()
-    cur.execute("SELECT username, password FROM users")
-    users = cur.fetchall()
-    con.close()
-    return users
-
